PO-2255 - code review updates

+++ src/main/java/uk/gov/hmcts/opal/service/report/FpRegisterService.java
@@ -28,7 +28,7 @@ public class FpRegisterService implements ReportInterface {
     public static class FpRegisterData implements ReportDataInterface {
 
         @Override
-        public int getNumberOfRecords() {
+        public long getNumberOfRecords() {
             return 0;
         }
